+++ nethsm/__init__.py
@@ -60,6 +60,7 @@ class State(enum.Enum):
     UNPROVISIONED = "Unprovisioned"
     LOCKED = "Locked"
     OPERATIONAL = "Operational"
+    FAILED = "Failed"
 
     @staticmethod
     def from_string(s: str) -> "State":
@@ -470,7 +471,7 @@ def _handle_exception(
     e: Exception,
     messages: Optional[dict[int, str]] = None,
     roles: Optional[list[Role]] = None,
-    state: Optional[State] = None,
+    state: Optional[Union[State, list[State]]] = None,
 ) -> NoReturn:
     from .client import ApiException
 
@@ -488,7 +489,7 @@ def _handle_api_exception(
     e: "ApiException[Any]",
     messages: Optional[dict[int, str]] = None,
     roles: Optional[list[Role]] = None,
-    state: Optional[State] = None,
+    state: Optional[Union[State, list[State]]] = None,
 ) -> NoReturn:
     from .client.api_response import ApiResponseWithoutDeserialization
     from .client.components.schema.error_response import ErrorResponseDict
@@ -514,7 +515,11 @@ def _handle_api_exception(
     elif e.status == 406:
         message = "Invalid content type requested"
     elif e.status == 412 and state:
-        message = f"Precondition failed -- this operation can only be used on a NetHSM in the state {state.value}"
+        if isinstance(state, list):
+            state_str = " or ".join([s.value for s in state])
+        else:
+            state_str = state.value
+        message = f"Precondition failed -- this operation can only be used on a NetHSM in the state {state_str}"
     elif e.status == 429:
         message = "Too many requests -- you may have tried the wrong credentials too often"
     else:
@@ -1647,7 +1652,9 @@ class NetHSM:
         try:
             self._get_api().system_reboot_post()
         except Exception as e:
-            _handle_exception(e, state=State.OPERATIONAL, roles=[Role.ADMINISTRATOR])
+            _handle_exception(
+                e, state=[State.OPERATIONAL, State.FAILED], roles=[Role.ADMINISTRATOR]
+            )
 
     def shutdown(self) -> None:
         try:
@@ -1662,7 +1669,9 @@ class NetHSM:
         try:
             self._get_api().system_factory_reset_post()
         except Exception as e:
-            _handle_exception(e, state=State.OPERATIONAL, roles=[Role.ADMINISTRATOR])
+            _handle_exception(
+                e, state=[State.OPERATIONAL, State.FAILED], roles=[Role.ADMINISTRATOR]
+            )
 
     def encrypt(
         self, key_id: str, data: Base64, mode: EncryptMode, iv: Optional[Base64] = None